backend/requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
websocket-client==1.7.0
requests==2.31.0
pyjwt==2.8.0
urllib3==2.1.0
certifi==2024.2.2
python-docx==0.8.11
Pillow==10.4.0
gunicorn==21.2.0
